Chat Filename Humanization + Larger Past Chats Box
1. Replaced all Date.now() instances with a humanized ISO8601 timestamp that is easily readable by users.

This has a seamless effect on the  embedded timestamping of chat mesages. Old chat files can still be edited (messages deleted), and continued.

Character PNG files get updated with the new date format, and this will be reflected in github but there is no functional change to the user.

2. Added character name to the front of all newly saved/imported chat files. Old chats will retain their original filename unless re-imported.

With these two changes, the new chat filenames are as such:

"Aqua - 2023-3-3 @03h 08m 36s 948ms.jsonl"

(I would like to make it smaller, perhaps by removing the milliseconds. Let me know if this is possible, or if it would introduce potential overwrites if users make/update files within the same second.)

3. Increased the size of the 'view past chats' popup box to max 800px so accommodate the new larger filenames.