Expressions: Add sprite aliasing for single chats
Duplicate characters get handled by the system, so why not sprites?
If a sprite is duplicated, users can specify the folder name that
the changed sprites are located in.

This commit only adds support for single person chats and breaks
compatibility with group chats. Another commit is required to fix
group sprite functionality.

Signed-off-by: kingbri <bdashore3@proton.me>