Script: Don't save character on chat creation
There's no need to save a character when a new chat is created.
The main reason for this commit is that after creating a new chat
with metadata, it immediately gets cleared after a few seconds.
This may be a race condition, but there's also no need for this
function call in the first place.

Fixes: Author's note not injecting on first message when new
chat is spawned.

Signed-off-by: kingbri <bdashore3@proton.me>