Generate: Fix appends on continue
Continue was in a semi-working state. However, the generated prompt
by continue was a mess and didn't remove essential parts such as
character name and prompt bias. This caused duplication and bad
generations.

Now, append the prompt bias after the CFG has been appended and then
clean up the continued cache before adding it to the final prompt.

Signed-off-by: kingbri <bdashore3@proton.me>