Generate: Migrate to array-based message concatenation
String addition is very limited in how flexible it can be. Using
an array will instead allow for various ways to shift around extension
prompts and insertion depths.

To preserve message order, each mesSend object contains both a message
and an array of extension prompts which are later added on top of
the message itself.

Signed-off-by: kingbri <bdashore3@proton.me>