World Info: Add character locking for entries
With the addition of depth, world info has become more important than
ever. However, there is a conflict that can occur when using multiple
world info or lorebooks on a character. Add the ability to lock
entries to a specific character (or characters) if needed.

This allows for adequate separation especially when there are character
cards each with their separate lorebooks that build off of a core lorebook.

Signed-off-by: kingbri <bdashore3@proton.me>