Translating to Russian language. Translation of the User Settings tab. (again)
Made a translation of the User Settings tab. I tried to translate everything that was not translated and tried to correct the index.html file in some places so that, for example, the PC and phone icons would not disappear during translation. Next, I'll try to translate other tabs if I'm not too lazy.
By the way, I wanted to ask, how can I translate the pop-up message in these buttons? I just asked once and they wrote to me that they are a different module altogether. (see photo)