Update server.js
Print warning if basicAuth username or password fails to parse.

In a normal case the user has no way to be informed if the username or password fails to parse. While this might end up being a skill issue on the users side it could help them to troubleshoot the issue.