Added option to delete chat backups when main chat and Character deleted
Added an option to delete the backups when the main chat file is deleted and all chat backups for a character when the character is deleted.

New flag in config.yaml "deleteChatBackupWithMain: false"

Added to \default\content\settings.json

Modified delete function in src/endpoints/chats.js to read the new config file and if true, search for all backups with a matching chat_id_hash to be deleted as well.