feature: essential strategy, ignoring budget caps
The budget cap default at 25% is a big gotcha at the start of chats, as a person may assume that a constant lore book entry will always be present, no matter what. This pull request adds a new 'essential' strategy which ignores the budget cap and always includes the lore book entry. It also adds clarification to the existing 'constant' entry.